requirements.txt
Django==1.11.29
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
PyJWT==2.8.0
django-filter
google-generativeai